fix: rename modal to model package
