+++ src/main/java/com/techconative/posf/model/POSFCriteria.java
@@ -1,4 +1,4 @@
-package com.techconative.posf.modal;
+package com.techconative.posf.model;
 
 import lombok.Getter;
 import lombok.Setter;
