+++ src/main/java/com/techconative/posf/util/POSFUtil.java
@@ -2,9 +2,9 @@ package com.techconative.posf.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.techconative.posf.modal.OrderingCriteria;
-import com.techconative.posf.modal.PageResult;
-import com.techconative.posf.modal.POSFCriteria;
+import com.techconative.posf.model.OrderingCriteria;
+import com.techconative.posf.model.PageResult;
+import com.techconative.posf.model.POSFCriteria;
 import exception.InvalidCriteriaException;
 import java.util.*;
 import java.util.regex.Pattern;
