+++ src/test/java/com/techconative/posf/model/DailyFields.java
@@ -1,4 +1,4 @@
-package com.techconative.posf.modal;
+package com.techconative.posf.model;
 
 import lombok.Data;
